Fix FCM registration for apps using firebase >= 20.1.1
With version 20.1.1 the Firebase Cloud Messaging SDK started to use
the Firebase Installations SDK, which affects the FCM registration
process.

The implementation of FCM registration in microG failed to pass extra
parameters that became relevant with the introduction of the
Firebase Installations SDK to the FCM registration endpoint.

These additional parameters are passed through to the endpoint with an 'X-' prefix.